Move LoadedAggregateTests to ModelTests in core lib. These tests no longer depend on LoadedAggregate.
